🌱 fix: split RBACDrillDown and ConfigMapDrillDown
Extract hooks and UI sub-components following the established
X.tsx + X.parts.tsx + useX.ts pattern.

RBACDrillDown.tsx: 465 → 137 lines, 18 → 4 hook calls
ConfigMapDrillDown.tsx: 477 → 169 lines, 13 → 6 hook calls

Closes #21873, #21877, #21806
