+++ web/src/components/drilldown/views/RBACDrillDown.tsx
@@ -1,47 +1,16 @@
-import { useState, useEffect, useRef, useCallback } from 'react'
-import { useLocalAgent } from '../../../hooks/useLocalAgent'
-import { useDrillDownWebSocket } from '../../../hooks/useDrillDownWebSocket'
+import { useState } from 'react'
 import { useDrillDownActions } from '../../../hooks/useDrillDown'
-import { ClusterBadge } from '../../ui/ClusterBadge'
-import { FileText, Code, Info, Loader2, Copy, Check, Server, Shield, ShieldCheck, User, RefreshCw, AlertTriangle } from 'lucide-react'
+import { FileText, Code, Shield, Info } from 'lucide-react'
 import { cn } from '../../../lib/cn'
-import { UI_FEEDBACK_TIMEOUT_MS } from '../../../lib/constants/network'
 import { useTranslation } from 'react-i18next'
-import { copyToClipboard } from '../../../lib/clipboard'
-
-// ---------------------------------------------------------------------------
-// Named constants — no magic numbers
-// ---------------------------------------------------------------------------
-
-/**
- * Maximum number of bindings the Describe and YAML tabs render inline.
- * Anything above this is truncated; a notice is displayed to the user
- * (Issue 9267).
- */
-const MAX_BINDINGS_TO_DESCRIBE = 10
+import { useRBACDrillDown, type DrillDownKind } from './useRBACDrillDown'
+import { RBACDrillDownHeader, RBACOverviewTab, RBACOutputPane } from './RBACDrillDown.parts'
 
 interface Props {
   data: Record<string, unknown>
 }
 
-interface RoleBinding {
-  kind: string
-  name: string
-  namespace?: string
-  roleName: string
-  roleKind: string
-}
-
 type TabType = 'overview' | 'describe' | 'yaml'
-
-/**
- * RBAC subject kinds that a *subject* column can hold. When the drilldown
- * is opened on a Role or RoleBinding row itself, the filter logic needs to
- * match by the role name in `roleRef`, not by the subject name (Issue 9264).
- */
-type DrillDownKind = 'User' | 'Group' | 'ServiceAccount' | 'Role' | 'RoleBinding' | 'ClusterRole' | 'ClusterRoleBinding'
-
-const SUBJECT_KINDS = new Set<DrillDownKind>(['User', 'Group', 'ServiceAccount'])
 
 export function RBACDrillDown({ data }: Props) {
   const { t } = useTranslation()
@@ -49,191 +18,29 @@ export function RBACDrillDown({ data }: Props) {
   const namespace = data.namespace as string | undefined
   const subject = data.subject as string
   const subjectType = ((data.type as string) || 'User') as DrillDownKind
-  const { isConnected: agentConnected } = useLocalAgent()
   const { drillToCluster, drillToNamespace } = useDrillDownActions()
-  const { runKubectl } = useDrillDownWebSocket(cluster)
 
   const [activeTab, setActiveTab] = useState<TabType>('overview')
-  const [clusterBindings, setClusterBindings] = useState<RoleBinding[]>([])
-  const [roleBindings, setRoleBindings] = useState<RoleBinding[]>([])
-  const [loading, setLoading] = useState(true)
-  const [loadError, setLoadError] = useState<string | null>(null)
-  const [describeOutput, setDescribeOutput] = useState<string | null>(null)
-  const [describeLoading, setDescribeLoading] = useState(false)
-  const [yamlOutput, setYamlOutput] = useState<string | null>(null)
-  const [yamlLoading, setYamlLoading] = useState(false)
-  const [copiedField, setCopiedField] = useState<string | null>(null)
-  const [refreshing, setRefreshing] = useState(false)
-  const mountedRef = useRef(true)
-  const copyTimerRef = useRef<ReturnType<typeof setTimeout>>(undefined)
 
-  useEffect(() => {
-    return () => {
-      mountedRef.current = false
-      clearTimeout(copyTimerRef.current)
-    }
-  }, [])
-
-
-  /**
-   * Returns true if the given binding matches the drilled-into target.
-   * - For User/Group/ServiceAccount, match by subject kind + name (the
-   *   binding references the drilled-into subject).
-   * - For Role / ClusterRole, match by roleRef.name (any binding that
-   *   references the drilled-into role).
-   * - For RoleBinding / ClusterRoleBinding, match by metadata.name (the
-   *   binding itself is the target).
-   *
-   * Previously the code only handled the User/Group/ServiceAccount case,
-   * so drilldowns opened from Role / RoleBinding rows always returned
-   * empty — Issue 9264.
-   */
-  const parseBindings = useCallback((json: string, kind: string): RoleBinding[] => {
-    try {
-      const parsed = JSON.parse(json)
-      const items = parsed.items || []
-      return items.filter((item: Record<string, unknown>) => {
-        if (SUBJECT_KINDS.has(subjectType)) {
-          const subjects = item.subjects as Array<{ kind: string; name: string }> | undefined
-          return subjects?.some(s => s.name === subject && s.kind === subjectType)
-        }
-        if (subjectType === 'Role' || subjectType === 'ClusterRole') {
-          const roleRef = item.roleRef as { kind: string; name: string } | undefined
-          return roleRef?.name === subject
-        }
-        if (subjectType === 'RoleBinding' || subjectType === 'ClusterRoleBinding') {
-          const meta = item.metadata as { name: string } | undefined
-          return meta?.name === subject
-        }
-        return false
-      }).map((item: Record<string, unknown>) => {
-        const roleRef = item.roleRef as { kind: string; name: string }
-        const meta = item.metadata as { name: string; namespace?: string }
-        return {
-          kind,
-          name: meta.name,
-          namespace: meta.namespace,
-          roleName: roleRef.name,
-          roleKind: roleRef.kind,
-        }
-      })
-    } catch {
-      return []
-    }
-  }, [subject, subjectType])
-
-  /**
-   * Fetch bindings from the cluster. Also clears the stale Describe/YAML
-   * output so the next tab switch re-fetches (Issue 9267).
-   */
-  const fetchBindings = useCallback(async () => {
-    if (!agentConnected) return
-    setLoading(true)
-    setLoadError(null)
-
-    try {
-      const [crbOut, rbOut] = await Promise.all([
-        runKubectl(['get', 'clusterrolebindings', '-o', 'json']),
-        namespace
-          ? runKubectl(['get', 'rolebindings', '-n', namespace, '-o', 'json'])
-          : runKubectl(['get', 'rolebindings', '--all-namespaces', '-o', 'json']),
-      ])
-
-      if (!mountedRef.current) return
-      setClusterBindings(parseBindings(crbOut, 'ClusterRoleBinding'))
-      setRoleBindings(parseBindings(rbOut, 'RoleBinding'))
-    } catch (err) {
-      if (!mountedRef.current) return
-      const errMsg = err instanceof Error ? err.message : 'Failed to fetch bindings'
-      setLoadError(errMsg)
-      setClusterBindings([])
-      setRoleBindings([])
-    }
-    setLoading(false)
-  }, [agentConnected, namespace, parseBindings, runKubectl])
-
-  const fetchDescribe = useCallback(async () => {
-    if (!agentConnected || describeOutput) return
-    setDescribeLoading(true)
-    const bindings = [...clusterBindings, ...roleBindings]
-    const parts: string[] = []
-    for (const b of bindings.slice(0, MAX_BINDINGS_TO_DESCRIBE)) {
-      if (!mountedRef.current) return
-      const args = b.kind === 'ClusterRoleBinding'
-        ? ['describe', 'clusterrolebinding', b.name]
-        : ['describe', 'rolebinding', b.name, '-n', b.namespace || 'default']
-      const out = await runKubectl(args)
-      if (out) parts.push(out)
-    }
-    if (!mountedRef.current) return
-    setDescribeOutput(parts.join('\n---\n') || t('drilldown.empty.noBindingsFound'))
-    setDescribeLoading(false)
-  }, [agentConnected, clusterBindings, describeOutput, roleBindings, runKubectl, t])
-
-  const fetchYaml = useCallback(async () => {
-    if (!agentConnected || yamlOutput) return
-    setYamlLoading(true)
-    const bindings = [...clusterBindings, ...roleBindings]
-    const parts: string[] = []
-    for (const b of bindings.slice(0, MAX_BINDINGS_TO_DESCRIBE)) {
-      if (!mountedRef.current) return
-      const args = b.kind === 'ClusterRoleBinding'
-        ? ['get', 'clusterrolebinding', b.name, '-o', 'yaml']
-        : ['get', 'rolebinding', b.name, '-n', b.namespace || 'default', '-o', 'yaml']
-      const out = await runKubectl(args)
-      if (out) parts.push(out)
-    }
-    if (!mountedRef.current) return
-    setYamlOutput(parts.join('\n---\n') || t('drilldown.empty.noBindingsFound'))
-    setYamlLoading(false)
-  }, [agentConnected, clusterBindings, roleBindings, runKubectl, yamlOutput, t])
-
-  const hasLoadedRef = useRef(false)
-
-  // Track agent connection state so a disconnect → reconnect also triggers a
-  // fresh fetch (Issue 9267). Without this the drilldown stayed blank after
-  // the agent dropped and returned.
-  const prevAgentConnectedRef = useRef(agentConnected)
-
-  useEffect(() => {
-    if (agentConnected && !hasLoadedRef.current) {
-      hasLoadedRef.current = true
-      fetchBindings()
-    }
-    // If the agent reconnected after having been disconnected, force a refresh
-    if (agentConnected && !prevAgentConnectedRef.current && hasLoadedRef.current) {
-      setDescribeOutput(null)
-      setYamlOutput(null)
-      fetchBindings()
-    }
-    prevAgentConnectedRef.current = agentConnected
-  }, [agentConnected, fetchBindings])
-
-  /**
-   * Manual refresh — re-fetches bindings and clears the Describe/YAML output
-   * so the next tab switch re-fetches fresh output (Issue 9267).
-   */
-  const handleRefresh = useCallback(async () => {
-    if (!agentConnected || refreshing) return
-    setRefreshing(true)
-    setDescribeOutput(null)
-    setYamlOutput(null)
-    try {
-      await fetchBindings()
-    } finally {
-      setRefreshing(false)
-    }
-  }, [agentConnected, fetchBindings, refreshing])
-
-  const handleCopy = (field: string, value: string) => {
-    copyToClipboard(value)
-    setCopiedField(field)
-    clearTimeout(copyTimerRef.current)
-    copyTimerRef.current = setTimeout(() => setCopiedField(null), UI_FEEDBACK_TIMEOUT_MS)
-  }
-
-  const totalBindings = clusterBindings.length + roleBindings.length
-  const hiddenBindingCount = Math.max(0, totalBindings - MAX_BINDINGS_TO_DESCRIBE)
+  const {
+    agentConnected,
+    clusterBindings,
+    roleBindings,
+    loading,
+    loadError,
+    describeOutput,
+    describeLoading,
+    yamlOutput,
+    yamlLoading,
+    copiedField,
+    refreshing,
+    totalBindings,
+    hiddenBindingCount,
+    fetchDescribe,
+    fetchYaml,
+    handleRefresh,
+    handleCopy,
+  } = useRBACDrillDown(cluster, namespace, subject, subjectType)
 
   const TABS: { id: TabType; label: string; icon: typeof Info }[] = [
     { id: 'overview', label: t('drilldown.rbac.bindingsTab', { count: totalBindings }), icon: Shield },
@@ -243,45 +50,17 @@ export function RBACDrillDown({ data }: Props) {
 
   return (
     <div className="flex flex-col h-full -m-6">
-      {/* Header */}
-      <div className="px-6 pt-6 pb-4">
-        <div className="flex items-center gap-6 text-sm">
-          <div className="flex items-center gap-2 px-3 py-1.5 rounded-lg bg-purple-500/10 border border-purple-500/20">
-            <User className="w-4 h-4 text-purple-400" />
-            <span className="text-muted-foreground">{subjectType}</span>
-            <span className="font-mono text-purple-400">{subject}</span>
-          </div>
-          {namespace && (
-            <button
-              onClick={() => drillToNamespace(cluster, namespace)}
-              className="flex items-center gap-2 hover:bg-purple-500/10 border border-transparent hover:border-purple-500/30 px-3 py-1.5 rounded-lg transition-all group cursor-pointer"
-            >
-              <span className="text-muted-foreground">{t('drilldown.fields.namespace')}</span>
-              <span className="font-mono text-purple-400 group-hover:text-purple-300">{namespace}</span>
-            </button>
-          )}
-          <button
-            onClick={() => drillToCluster(cluster)}
-            className="flex items-center gap-2 hover:bg-blue-500/10 border border-transparent hover:border-blue-500/30 px-3 py-1.5 rounded-lg transition-all group cursor-pointer"
-          >
-            <Server className="w-4 h-4 text-blue-400" />
-            <span className="text-muted-foreground">{t('drilldown.fields.cluster')}</span>
-            <ClusterBadge cluster={cluster.split('/').pop() || cluster} size="sm" />
-          </button>
-          {/* Refresh button (Issue 9267) — re-fetches bindings and clears
-              stale Describe/YAML output so the next tab switch re-runs them. */}
-          <button
-            onClick={handleRefresh}
-            disabled={!agentConnected || refreshing}
-            className="ml-auto flex items-center gap-2 px-3 py-1.5 rounded-lg text-sm text-muted-foreground hover:text-foreground hover:bg-secondary/50 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
-            data-testid="rbac-drilldown-refresh"
-            aria-label={t('common.refresh')}
-          >
-            <RefreshCw className={cn('w-4 h-4', refreshing && 'animate-spin')} />
-            <span>{t('common.refresh')}</span>
-          </button>
-        </div>
-      </div>
+      <RBACDrillDownHeader
+        subjectType={subjectType}
+        subject={subject}
+        namespace={namespace}
+        cluster={cluster}
+        agentConnected={agentConnected}
+        refreshing={refreshing}
+        onDrillToNamespace={drillToNamespace}
+        onDrillToCluster={drillToCluster}
+        onRefresh={handleRefresh}
+      />
 
       {/* Tabs */}
       <div className="border-b border-border px-6">
@@ -314,149 +93,42 @@ export function RBACDrillDown({ data }: Props) {
       {/* Tab Content */}
       <div className="flex-1 overflow-y-auto p-6 space-y-6">
         {activeTab === 'overview' && (
-          <div className="space-y-6">
-            {loadError && (
-              <div className="flex items-center gap-2 p-3 rounded-lg bg-red-500/10 border border-red-500/30 text-red-300">
-                <AlertTriangle className="w-4 h-4 shrink-0" />
-                <span className="text-sm">{loadError}</span>
-              </div>
-            )}
-            {loading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-primary" />
-                <span className="ml-2 text-muted-foreground">{t('drilldown.rbac.loadingBindings')}</span>
-              </div>
-            ) : totalBindings === 0 ? (
-              <div className="p-4 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-center">
-                <p className="text-yellow-400">{t('drilldown.rbac.noBindingsForSubject', { type: subjectType, subject })}</p>
-              </div>
-            ) : (
-              <>
-                {clusterBindings.length > 0 && (
-                  <div>
-                    <h3 className="text-sm font-semibold text-foreground mb-3 flex items-center gap-2">
-                      <ShieldCheck className="w-4 h-4 text-green-400" />
-                      {t('drilldown.rbac.clusterRoleBindingsHeader', { count: clusterBindings.length })}
-                    </h3>
-                    <div className="space-y-2">
-                      {clusterBindings.map((b) => (
-                        <div key={b.name} className="p-3 rounded-lg bg-green-500/10 border border-green-500/20 flex items-center justify-between">
-                          <div>
-                            <div className="font-mono text-sm text-green-400">{b.name}</div>
-                            <div className="text-xs text-muted-foreground mt-1">
-                              {b.roleKind}: <span className="text-foreground">{b.roleName}</span>
-                            </div>
-                          </div>
-                          <span className="text-xs px-2 py-1 rounded bg-green-500/10 text-green-400 border border-green-500/20">{t('drilldown.rbac.clusterWide')}</span>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-
-                {roleBindings.length > 0 && (
-                  <div>
-                    <h3 className="text-sm font-semibold text-foreground mb-3 flex items-center gap-2">
-                      <Shield className="w-4 h-4 text-blue-400" />
-                      {t('drilldown.rbac.roleBindingsHeader', { count: roleBindings.length })}
-                    </h3>
-                    <div className="space-y-2">
-                      {roleBindings.map((b) => (
-                        <div key={`${b.namespace}-${b.name}`} className="p-3 rounded-lg bg-blue-500/10 border border-blue-500/20 flex items-center justify-between">
-                          <div>
-                            <div className="font-mono text-sm text-blue-400">{b.name}</div>
-                            <div className="text-xs text-muted-foreground mt-1">
-                              {b.roleKind}: <span className="text-foreground">{b.roleName}</span>
-                              {b.namespace && <> {t('drilldown.rbac.inNamespace')} <span className="text-foreground">{b.namespace}</span></>}
-                            </div>
-                          </div>
-                          {b.namespace && (
-                            <span className="text-xs px-2 py-1 rounded bg-blue-500/10 text-blue-400 border border-blue-500/20">{b.namespace}</span>
-                          )}
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-              </>
-            )}
-          </div>
+          <RBACOverviewTab
+            loadError={loadError}
+            loading={loading}
+            totalBindings={totalBindings}
+            subjectType={subjectType}
+            subject={subject}
+            clusterBindings={clusterBindings}
+            roleBindings={roleBindings}
+          />
         )}
-
         {activeTab === 'describe' && (
           <div>
-            {describeLoading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-primary" />
-                <span className="ml-2 text-muted-foreground">{t('drilldown.status.runningDescribe')}</span>
-              </div>
-            ) : describeOutput ? (
-              <div className="relative">
-                <button
-                  onClick={() => handleCopy('describe', describeOutput)}
-                  className="absolute top-2 right-2 px-2 py-1 rounded bg-secondary/50 text-xs text-muted-foreground hover:text-foreground flex items-center gap-1"
-                >
-                  {copiedField === 'describe' ? <><Check className="w-3 h-3 text-green-400" /> {t('common.copied')}</> : <><Copy className="w-3 h-3" /> {t('common.copy')}</>}
-                </button>
-                {/* Truncation notice (Issue 9267) */}
-                {hiddenBindingCount > 0 && (
-                  <div
-                    className="mb-2 px-3 py-2 rounded-md bg-yellow-500/10 border border-yellow-500/20 text-xs text-yellow-400"
-                    data-testid="rbac-drilldown-truncation-notice"
-                  >
-                    {t('drilldown.rbac.truncationNotice', {
-                      shown: MAX_BINDINGS_TO_DESCRIBE,
-                      total: totalBindings,
-                      hidden: hiddenBindingCount })}
-                  </div>
-                )}
-                <pre className="p-4 rounded-lg bg-black/50 border border-border overflow-auto max-h-[60vh] text-xs text-foreground font-mono whitespace-pre-wrap">
-                  {describeOutput}
-                </pre>
-              </div>
-            ) : (
-              <div className="p-4 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-center">
-                <p className="text-yellow-400">{t('drilldown.empty.localAgentNotConnected')}</p>
-              </div>
-            )}
+            <RBACOutputPane
+              loading={describeLoading}
+              output={describeOutput}
+              copiedField={copiedField}
+              copiedFieldKey="describe"
+              hiddenBindingCount={hiddenBindingCount}
+              totalBindings={totalBindings}
+              loadingText={t('drilldown.status.runningDescribe')}
+              onCopy={handleCopy}
+            />
           </div>
         )}
-
         {activeTab === 'yaml' && (
           <div>
-            {yamlLoading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-primary" />
-                <span className="ml-2 text-muted-foreground">{t('drilldown.status.fetchingYaml')}</span>
-              </div>
-            ) : yamlOutput ? (
-              <div className="relative">
-                <button
-                  onClick={() => handleCopy('yaml', yamlOutput)}
-                  className="absolute top-2 right-2 px-2 py-1 rounded bg-secondary/50 text-xs text-muted-foreground hover:text-foreground flex items-center gap-1"
-                >
-                  {copiedField === 'yaml' ? <><Check className="w-3 h-3 text-green-400" /> {t('common.copied')}</> : <><Copy className="w-3 h-3" /> {t('common.copy')}</>}
-                </button>
-                {/* Truncation notice (Issue 9267) */}
-                {hiddenBindingCount > 0 && (
-                  <div
-                    className="mb-2 px-3 py-2 rounded-md bg-yellow-500/10 border border-yellow-500/20 text-xs text-yellow-400"
-                  >
-                    {t('drilldown.rbac.truncationNotice', {
-                      shown: MAX_BINDINGS_TO_DESCRIBE,
-                      total: totalBindings,
-                      hidden: hiddenBindingCount })}
-                  </div>
-                )}
-                <pre className="p-4 rounded-lg bg-black/50 border border-border overflow-auto max-h-[60vh] text-xs text-foreground font-mono whitespace-pre-wrap">
-                  {yamlOutput}
-                </pre>
-              </div>
-            ) : (
-              <div className="p-4 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-center">
-                <p className="text-yellow-400">{t('drilldown.empty.localAgentNotConnected')}</p>
-              </div>
-            )}
+            <RBACOutputPane
+              loading={yamlLoading}
+              output={yamlOutput}
+              copiedField={copiedField}
+              copiedFieldKey="yaml"
+              hiddenBindingCount={hiddenBindingCount}
+              totalBindings={totalBindings}
+              loadingText={t('drilldown.status.fetchingYaml')}
+              onCopy={handleCopy}
+            />
           </div>
         )}
       </div>
